lec8-grid-search-concrete/requirements.txt
absl-py==2.2.2
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.58.0
gast==0.4.0
google-auth==2.40.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
joblib==1.5.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==1.23.5
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.1.3
scipy==1.15.3
six==1.17.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-macos==2.10.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
ucimlrepo==0.0.7
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2